+++ packages/ui/package.json
@@ -16,6 +16,7 @@
   "dependencies": {
     "@radix-ui/react-label": "^2.1.7",
     "@radix-ui/react-slot": "^1.2.3",
+    "@radix-ui/react-tooltip": "1.2.7",
     "@tanstack/react-query": "5.81.5",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
